pyproject.toml
[tool.poetry]
name = "myswarm"
version = "0.1.0"
description = ""
authors = ["rycharlind <rycharlind@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
swarm = {git = "https://github.com/openai/swarm.git"}
python-dotenv = "^1.0.1"
google-api-python-client = "^2.151.0"
firecrawl-py = "^1.4.0"
openai = "^1.53.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
